Changes in the last commit.
I had removed the code when we first time installs the Addon then radio
button for the lead will not be selected. But if lead will not be
selected then text field was not displaying. So the radio button for
Lead will be selected for new installation of the addon.