Salesforce Setting name & Other description
We have added 8 settings in the Salesforce addon. Named as 1) Export abandoned cart data to Salesforce CRM 2) Automatically add abandoned cart data to Salesforce CRM 3) Automatically add abandoned cart data to salesforce CRM after set time 4) Salesforce Username 5) Salesforce Password 6) Salesforce Security Token 7) Create abandoned carts as 8) Company name of lead

 Here, first 3 settings are same as the Agile CRM. Enable the salesforce CRM, cron time & enable automatic cron setting.

Other settings: 1) Username- It will require you to create the DEVELOPER account on the salesforce. And provide that username in the setting 2) password of the created developer account. 3) Security token: How do we get the security token is described in the setting description. Please follow that.

4) How to create abandoned cart - Here admin can choose how create cart? As a LEAD or CUSTOMER. When you create cart as a LEAD it will show new setting for the COMPANY name.

NOTE - when the lead is enabled notes will be not added, instead of that the only one-time description will be added to that lead.

Note - when we create the contact it will add the abandoned cart data in the notes.

Automatic cron will not add the data again if it is already exported to CRM

NOTE- when LEAD is created the company name is required field. If the setting is blank then we are adding default value: 'Abandoned Cart Plugin' for it